Implement GetBestString and modify FetchParam to make sure fetched string[0] is not empty if any other string index has a useful string
These are used to make it easier for script commands to access string literals and string variables. Now that the frontend always loads English as language 0, the Japanese game scripts have all their strings as language 1, and the script engine needs to be able to easily grab the relevant strings without worrying too much about the active language.