Sakufiber: jumps to first out-of-bound code byte should be accepted
Fixes a case where the last part of a label's code has an if-then etc, and failing the condition causes a jump to the next byte after the end of the if-block, which is out of bounds. This will now be treated as just a normal slide into the next chained label.