decomp: Use the new mcfileio unit for file loading, use exceptions for decomp errors
This work isn't quite complete. It compiles and runs about as well as before, except
there's an obvious memory leak in the dispatcher, and results aren't collected or
reported at all. With a bit of polish, the code ought to be cleaner and easier to
work with. The dispatcher just needs a redesign, perhaps with extra fuzziness so it
can autodetect most filetypes...